remove private network TODO
I added that TODO based on a misunderstanding of the function of the
private network between the Ganeti nodes. It's not there to provide a
private network between *instances*, but really just the nodes, for
DRBD and Ganeti traffic.

So it makes no sense to allocate private IPs to the instances in that context.