ring the bell in terminals when a step is completed
The rationale of this is that we want to speed up the upgrade process
and not let the upgrade wait for operator wandering away in another
distraction. Inversely, it allows operators to exactly do other things
instead of just staring at the terminal, waiting for the process to
complete.

This assumes users either like bell noises or have turned them off in
their terminal. In my case, the terminal is set to send a "urgent"
signal to the window manager when it receives a "bell", using this
Xresources configuration:

    URxvt.urgentOnBell:  true
    XTerm*bellIsUrgent:  true
    XTerm*visualBell: false

In i3 with i3status (and py3status), the affected workspace gets a
nice red background, telling me the upgrade is waiting for my input.