get rid of install-disks script in favor of setup-storage
The thing just works. It does everything we need except maybe picking
arbitrary labels for RAID arrays (and that might even be possible to
work around).

By using a configuration file and setup-storage, we bring ourselves
closer to standardizing and automating the installer.

setup-storage also has the advantage of being idempotent, mostly: I
was able to run and rerun it multiple times from the rescue shell. The
only thing I had to do was to close LUKS devices, if I remember
correctly. But it was much easier to test than a shell script.

Finally, it provides nice features like dumping the fstab, crypttab
and mdadm.conf for us to use. We don't use those just yet.

It also provides a shell script that can be sourced to get various
targets, which we use.