specify the UUID correctly
we could have also used the new swap UUID and edited the fstab, but
that would have required more steps:

 1. mount the filesystem
 2. edit fstab
 3. unmount

... and another thing to redo after the last resync.

By enforcing a UUID from the start, we don't need to edit the fstab at
all. I can't think of a compelling reson why reusing the UUID might be
a problem.