trivial disk replacement procedure
While this procedure is available in a billion places already, this is
the kind of stuff that you mess up when you get awaken at 3AM and can
barely type.

Let's see if we can get the human error out of the equation a little
more by helping those poor infected water bags a little.