merge the firewall poking and CA generation in bootstrap (#32914)
The rationale here is we remove one more step in the installer:
instead of going back and forth between the master and client, we run
one command on both, simultaneously, and only transfer the SHA256
checksum between the two.

This could eventually be automated in such a way that we can do this
unattended if we have SSH access on both machines.